requirements.txt
pdfminer.six
PyPDF2
python-docx
spacy
nltk
Flask
